requirements.txt
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.4.0
django-polymorphic==3.1.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
numpy==1.26.4
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
rest-framework-simplejwt==0.0.2
setuptools==68.2.2
sqlparse==0.5.0
validate-email==1.3
wheel==0.41.2
